# spring-rest-assured-cucumber-testcontainer/src/main/resources/api/api.yaml
openapi: 3.0.1
info:
  title: Jobs API
  description: This is a simple API Application created to demonstrate the capabilities
    of Rest Assured for API Testing
  contact:
    name: Rakesh
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Job
    description: The Job API
paths:
  /api/v1/jobs/{jobId}:
    get:
      tags:
        - Job
      summary: Find Job post by id
      operationId: findById
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobResponse'
    put:
      tags:
        - Job
      summary: Update a Job Post
      operationId: updateJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
        required: true
      responses:
        "204":
          description: No Content
    delete:
      tags:
        - Job
      summary: Delete a Job post
      operationId: deleteJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
  /api/v1/jobs:
    get:
      tags:
        - Job
      summary: Find all available Job posts
      operationId: findAllJobs
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedContentJobResponse'
    post:
      tags:
        - Job
      summary: Add a new Job post
      operationId: addNewJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobResponse'
components:
  schemas:
    JobRequest:
      required:
        - description
        - jobTitle
        - jobType
      type: object
      properties:
        jobTitle:
          type: string
          example: Senior Backend Engineer (Java)
        description:
          type: string
          example: Brief description of the Job.
        jobType:
          type: string
          enum:
            - FULL_TIME
            - CONTRACTOR
            - PART_TIME
        datePosted:
          type: string
          format: date-time
        jobLink:
          pattern: "[(http(s)?):\\/\\/(www\\.)?a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,6}\\\
            b([-a-zA-Z0-9@:%_\\+.~#?&//=]*)"
          type: string
          example: https://mycompany.breezy.hr/p/22b17727d5f4-senior-backend-engineer-java
    JobResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10001
        jobTitle:
          type: string
          example: Senior Backend Engineer (Java)
        description:
          type: string
          example: Brief description of the Job.
        jobType:
          type: string
          enum:
            - FULL_TIME
            - CONTRACTOR
            - PART_TIME
        datePosted:
          type: string
          format: date-time
        jobLink:
          type: string
          example: https://mycompany.breezy.hr/p/22b17727d5f4-senior-backend-engineer-java
    PagedContentJobResponse:
      type: object
      properties:
        isFirst:
          type: boolean
        isLast:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/JobResponse'
        totalElements:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32